requirements.txt
click
cooler
graphviz
matplotlib
numpy
pandas <2.0
pybigwig
pydot
python >=3.9
scikit-learn
scipy
tensorflow-gpu  
tqdm